pyproject.toml
[tool.poetry]
name = "django-training"
version = "0.1.0"
description = ""
authors = ["Mohammed-Khedr <mohamed.khedr1714@gmail.com>"]
readme = "README.md"
packages = [{include = "django_training"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
django-environ = "^0.9.0"
Pillow = "^9.2.0"
django-imagekit = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"